Drop inode_data, instead inline in inode
The idea for this was to save the space for identical inodes.
However we need both the extra offset (8 bytes) for all inodes,
and sharing is not great. In my 4mb image descriptor the savings
was just 100k, which is imho not worth the extra complexity and
cost of double reads per inode.