Make xattr storage more compact
Instead of using a lot of indirection we make each list of xattrs
a single chunk of memory with all information. This means we can't
reuse the individual keys and values between inodes, however we
can reuse the entire xattrs combo, which is equally likely to be shared
in practice.

In fact, shared chunks are only useful if > sizeof(struct
lcfs_vdata_s) == 16. So the new serialization is likely more efficient
in practice with smaller keys/values. This is proved in my example
chroot which is 3832084 bytes with the old xattr format and 3831048
bytes with the new format. This is a meager saving of 1036 bytes, but
proves the point.

The savings is not the point here though, but the fact that all the
xattrs for an inode can now be read in one read operation.