dir: Store directories more compact
We store the names as well as the dentries in a single chunk allowing
for easier more efficient reads. Additionally this saves 200k for my 3MB
metadata file due to the use of less offsets.

We also store the d_type in the dentry so we can return it without
seeking to the target inode.