Change file format layout
This changes the main format of the descriptor. It is now
in two parts, the first is the inode data table which statrts
(with the root inode) directly after the header. This contains
an array lcfs_inode_s, each optionally followed by the payload of that
inode which is either lcfs_backing_s, lcfs_dir_s, or a symlink target.

The second part is a variable data accesses with vdata like before,
but now only used by xattrs. The size of the inode table is in the
header so we can know where the variable data starts.

With these changes the descriptor size for my testcase goes from 3.3Mb
to 3.0Mb.

Some other changes as part of this:

Inode order is more sane, its stored in breadth-first from the root.

The vdata are now indexed only by 32bit, as we're unlikely to need
more than 4GB of xattrs.

link targets are now stored without terminating zero.

We're better at canonicalizing empty payloads.

Writing the descriptor now doesn't need to build a huge vdata, because
we can stream the output of the inodes and payloads, so memory use is
better.