Use more compact form for inode table
This uses the low 8 bit (of 64) from the inode number as a flag for
which parts of the inode data is stored in the file, with the outstanding
parts being filled with default values. It also moves st_size to such
a flag-based part of the stat info, making regular and symbolic links
use the payload in the same way.

This significantly lowers the size of the images, as seen below for
the new test-assets:

Originally:

9.2M cs9-x86_64-developer
3.2M cs9-x86_64-minimal
18M f36-x86_64-silverblue

Now:
5.8M cs9-x86_64-developer
2.1M cs9-x86_64-minimal
12M f36-x86_64-silverblue

This encoding also fixes the issues with the timespec encoding
in #22.