Add support for fs-verity digests in the filesystem
This if you pass --compute-digests to writer it computes the
fs-verity sha25 digests and embeds in the descriptor for regular files.

The digests are dumped by `dump`, and if set verified by the kernel
to match with the backing file each time that is used.

This also changes the way the root flags works. We now encode the root
flags in the descriptor header. This is because we definately don't want
the digest flag set for the root.