lib: Use custom implementation of fsverity digest computation
This is a re-implememtation of the fsverity digest computation
that works like a traditional API, where you feed it bytes rather
than have it call a callback for more data.

This model will allow us to e.g. compute the digest of the
image without writing it to disk first (which we will want
for ostree).

It also means the library need not link to libfsverity, only
openssl. This might be useful for people depending on the
library.

Signed-off-by: Alexander Larsson <alexl@redhat.com>