Split out actual dentries to variable data
This moves the dentry chunks to the variable data section, instead of
being in the inode payload. The main reason for this is that we
can then easily handle the dentries, as well as the xattrs indirectly
via page mappings instead of reading from files, making sharing easier
and avoiding memory copies.

This commit also have some semi-related changes:
 * The new dentry has a name offset, allowing random seeks. The goal
   is to use this for binary search lookups.
 * The xattr variable data offset is now 64bit to ensure we don't
   accidentally run out of space for large filesystems, now that
   we have more variable data.
 * 4k align the vdata section

Signed-off-by: Alexander Larsson <alexl@redhat.com>