kernel: Convert cfs_dir_data to cfs_inode_data
This struct now contains all the information that cfs-reader.c kept
from the inode, which now includes the previous dir data as well as
the (validated) payload length.

This was previously stored separately and passed separately in addition to
dir_data to various cfs-reader.c functions that now only need one argument
for both of these.

Signed-off-by: Alexander Larsson <alexl@redhat.com>