Change when we require fs-verity
This changes when we verify fs-verity. The goal is to have it easy to
create fs-verity enabled images with `mkcomposefs --digest-store=`, but
also make it easy to consume them without requiring you to explicitly
disable verity.

The new setup is a single option verity_check=[0,1,2]. 0 means we
never require verity. 1 means we veritfy backing file fs-verity if the
image specifies a verity digest. 2 additionally means we require the image
to specify a digest for all non-empty files.

Fixes https://github.com/containers/composefs/issues/74

Signed-off-by: Alexander Larsson <alexl@redhat.com>