Drop chunking of directories and use larger buffers
This changes how directories are handled. The dirents for a directory
is now a separate vdata, which can now be larger than 4k (which now
also affects xattrs). This is made possible by using vm_map_ram()
for multi-page cfs_bufs.

The dirents doesn't have a flag, instead we just always store it in
the inode. This is in preparation for for further simplification where
we just always store all fields of the inode on disk.

Signed-off-by: Alexander Larsson <alexl@redhat.com>