kernel: Properly handle stacking
This makes composefs work both when the descriptor and when the
backing files are on a stacking filesystem (such as overlayfs), and
also makes stacking on top of composefs work.

For the later this implements d_real(), and to do this we need to be
able to return an owned reference to the "lower" dentry. In order to
do this we need need to hold a ref to it, so I added a lazily
initialized backing_dentry to the inode data which is filled in when
first needed.

For the first, we use d_real() to compute the real backing dentry/inode
for the descriptor and for the backing files. In addition to making the
overlay stacking work at all this also ensures that we're guaranteed to
do i/o on the same inode that we do the fs-verity digest validation on.

Signed-off-by: Alexander Larsson <alexl@redhat.com>