writer: Add support for cloning the tree while writing
This will be used later when we need to modify the tree while
writing it out. For example adding "." and ".." children, and changing
xattrs.

When we do this we also do a deep destroy of the tree which removes
all children first. This way we can handle the refcount loops that
can otherwise happen with "." and "..".

Signed-off-by: Alexander Larsson <alexl@redhat.com>