kernel: Avoid possibly unnecessary readhead
This switches cfs-reader to always use cfs_get_buf, rather than
sometimes using kernel_read(), as kernel_read() always uses readahead
heuristics that may not be suitable for the filesystem usecase. For
example it may readhead content of a subdirectory that is then not
used by userspace.

On some usecases this is a performance regression, but not really a
significant difference.

Signed-off-by: Alexander Larsson <alexl@redhat.com>