writer: Fix uninitialized read with xattrs
I was testing out composefs with `-fsanitize=address`.  I highly
recommend this as it's cheap and will absolutely find bugs.

In this case, what is happening here is that it's allowed for
an initially `listxattr` call to return a larger size than the *real*
size.  So we must take the second return value as the actual size.

Concretely in my setup, the initial `listxattr` call returns 40,
but the second one returns 17.  With the current code, we just walk
off the end of what was written by the Linux kernel and start
reading uninitialized heap memory, which is caught
by `-fsanitize=address`.

Signed-off-by: Colin Walters <walters@verbum.org>