Drop kernel code and references to it
It seems very unlikely atm that this is going to end up accepted
in the upstream kernel, so lets drop this code. If we don't the code
will just keep diverging from the upstream kernel over time.

If at some point we change our mind, the code is still available in the
git history.

Signed-off-by: Alexander Larsson <alexl@redhat.com>