composefs-from-json: Store mtimes in UTC
The time as stored in json includes a timzeone, but for the
erofs value we want to store the time in UTC. Unfortunately the
time apis are kind of lame here, timegm() always assumes the
data in the struct tm is in gmt time. So we have to extract the
gmt offset from the initial strptime call and apply it manually.

Signed-off-by: Alexander Larsson <alexl@redhat.com>