erofs: Support the new bloom filter erofs compat feature
This matches:
  https://git.kernel.org/pub/scm/linux/kernel/git/xiang/erofs.git/commit/?h=dev&id=fd73a4395d477ae134f319f7368a9f8a6264fd8b

We set the COMPAT_XATTR_FILTER flag on the superblock, and we add to
each xattr header a 32bit bloom filter which defaults to all bits
set, and then we compute a hash for each key modulo 32, and unset
that bit in the bloom filter.

Older kernels will ignore this, but newer kernels will use it
to make xattr lookup more efficient.

NOTE: This is a format-breaking change, so we also update the test
checksums. We should merge this before freezing the format.

Signed-off-by: Alexander Larsson <alexl@redhat.com>