lib: Add support for inlining small data files
For small files, it makes no sense to use a redirect to the basedir.
The size of the redirect xattr value itself is typically 65 bytes, and
on top of that comes the xattr name overhead. Things are even worse if
we also want to store the fs-verity digest. The overhead for a very
small file is significant.

So, we add the ability to store the file content to save in a
lcfs_node. And when serializing the erofs layer, if it is set we
skip the redirect xattrs, and just directly store the data in the
erofs file (instead of making it a sparse file).

Signed-off-by: Alexander Larsson <alexl@redhat.com>