lib: Add LCFS_BUILD_NO_INLINE option to build flags
Unless you pass this to lcfs_load_node_from_file() then it will
automatically add the file content for small files to the lcfs_nodes,
which means that the file data will be stored in the erofs image instead
of in a separate file in the basedir.

The limit for small files is 64 bytes, which is the size of a sha256
digest in hex form. That is typically the size of a redirect xattr
that would be used otherwise, so the end result is that we always get
a smaller image if we inline these. 64 is also the typical size of a
erofs inode entry, so this will inline well in the inode table.

Signed-off-by: Alexander Larsson <alexl@redhat.com>