Drop LCFS_MOUNT_FLAGS_DISABLE_VERITY
During upstream overlayfs work, the default for verity xattr was
changed to verity=off, so there is no need to specifically have a wey
to disable it. All that results in is mount error on kernels that
don't support verity=.

So, we just support ENABLE_VERITY which implies verity=require, which
is the thing we want anyway.

Signed-off-by: Alexander Larsson <alexl@redhat.com>