simplify mkcomposefs
Simplifies the mkcomposefs arguments to always compute digest and
by-digest payloads, dropping the --absolute option and related things,
as we couldn't think of a usecase for it right now. If we have a more
precise usecase later we can add it back in a way that makes sense for
the usecase.

Signed-off-by: Alexander Larsson <alexl@redhat.com>