tools: Don't user error.h for better portability
This is a glibc specific extension. We were already using the BSD
alternative err()/errx() in some places, and that seems more widely
suported, so lets just use that everywhere.

Most of the conversions are trivial:

 * error(EXIT_FAILURE, errno, => err(EXIT_FAILURE,
 * error(EXIT_FAILURE, 0, => errx(EXIT_FAILURE,
 * error(0, 0, => fprintf(stderr,
 * error(0, errno, => perror(

Although in some places we had to set errno to be able to use err().

Signed-off-by: Alexander Larsson <alexl@redhat.com>