Add mentioned licenses to the code. Change spec to use SPDX
As part of the Fedora package review, I ran scancode-toolkit to get
proper SPDX codes. In doing so I saw that various licenses were
referenced in the code but not included in the package. Those were

Apache-2.0 for composefs/tools/read-file.{c,h}
LGPL-3.0 for   composefs/libcomposefs/lcfs-fsverity.c
GPL-3.0 for composefs/tools/composefs-dump.c,
            composefs/tools/composefs-from-json.c,
            composefs/tools/composefs-info.c,
            composefs/tools/mkcomposefs.c,
            composefs/tools/mountcomposefs.c

There is some BSD 2 clause code included, but the code itself has the
complete license included in it.

Signed-off-by: Stephen Smoogen <ssmoogen@redhat.com>