lib: Use new mount api for overlay to mount overlayfs
This is required on kernel >= 6.5 to handle escaped commas in upper or
lowerdirs due to an overlayfs change in escape handling behaviour
caused by the switch to the new mount api internally.

For anything post 6.5 we fallback to the old legacy implementation
which supports escaping.

See https://github.com/containers/composefs/issues/213 for details.

Signed-off-by: Alexander Larsson <alexl@redhat.com>