Make errno handling cleaner with errint_t type
In some places we return an int which is -1 for error and errno set,
and in some places we return -errno for error. This adds a new
errint_t type to make it more clear what is happening. Also all users
are changed to use an errint_t err for the new type, and keep int res
for the old type.

Signed-off-by: Alexander Larsson <alexl@redhat.com>