mkcomposefs: Use sandboxing when --from-file is used
This helps protecting the parsing code if some external process
is feeding data to mkcomposefs to create an image.

Its not really useful to use a sandbox when a directory is specified
though, as it needs to do way to many different kinds of operations,
as well as reading actual uid/gid:s etc.

Signed-off-by: Alexander Larsson <alexl@redhat.com>