lib: Keep children sorted and lookup using binary search
We're hitting this child lookup a lot when inserting file due to the
check for existing files in add_child(), and then doubling this when
cloning the tree in the writer.

So, this makes lcfs_node_add_child() guarantee that children are
sorted always, which means we can use a binary seach for all lookups,
and we can avoid the qsort when canonicalizing images.

This speeds up mkcomposefs on a large --from-file image from 790 msec
to 470 msec.

Signed-off-by: Alexander Larsson <alexl@redhat.com>