writer: Limit to UINT16_MAX xattrs
This is the current EROFS limit, and we have a similar
check in the writer code.  We should not allow
internal memory representation to go above it.

There's a lot of potential followup work to change our usage of
size_t here to be uint16_t, but that can come later.

Signed-off-by: Colin Walters <walters@verbum.org>