Add new version 1 that uses the new xwhiteout marking approach
Linux upstream is making some change to how xwhiteout directories
are marked. They are now marked with an opaque xattr with content "x".

See:
 https://lore.kernel.org/linux-unionfs/CAOQ4uxh5x_-1j8HViCutVkghA1Uh-va+kJshCuvB+ep7WjmOFg@mail.gmail.com/T/#t

This creates a new image format version, 1, which adds these xattrs
(and the old ones for backwards compat with linux 6.7).

In addition, the tests are updated to handle this, and we create a new testcase that validates both v0 and v1 keeps working for whiteout files.

Note: This changes the output of the CLI tools to use the new version
by default, and you need to pass --verison-format=0 to get the old
version. This is ok at this point, as most images are unaffected by
the version due to lack of whiteouts, and not many people use
composefs yet.

Signed-off-by: Alexander Larsson <alexl@redhat.com>