Add support for max_version
Instead of specifying just the version we now specify the min and max
version, and composefs tries to keep the version as low as possible
but may bump it if needed for newer features. This is necessary as we
now record the version in the image, and we want to avoid
unnecessarily storing a later version in the image, as would change
the digest of images where this is not needed.

We also introduce new macros for the default max version to use in the
tooling, instead of just using the max version. This allows more
careful bumping of the default later.

Signed-off-by: Alexander Larsson <alexl@redhat.com>