mount: Use the new lowerdir+ and datadir+ options
These were added in Linux 6.7 and replace the old append mechanism
which now no longer work. See commit 24e16e385f for details of this.

Also, I added a check that we're actually using the real new mount api
rather than the legacy implementation of it, because otherwise we will
not see error reports during fsconfig, which this code relies on.

Signed-off-by: Alexander Larsson <alexl@redhat.com>