mount: Allow building when macro MOUNT_ATTR_IDMAP is not available
This is to allow building the software on machines not having the
MOUNT_ATTR_IDMAP macro in header "linux/mount.h". When that macro is not
available, the dependency on struct mount_attr is also eliminated (which
is good since both the macro and the struct were added to the kernel
uapi virtually at the same time).

With the changes in this commit, errors would be thrown at runtime when
mounting the erofs image, but only if the idmap feature is used; this
resembles the behavior when the "new mount API" is not detected.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>