tests: Fix extra `-` in should-fail tests
I had an extra `-` at the end here which was being interpreted
as an xattr...It turns out that `mkcomposefs` is liberal here
in what it accepts (string parsing in C stinks) and interprets
a missing `=` to be an xattr with the provided name, and
no value.

In this *specific* case that's an xattr with the name `-`.

The Rust parser is stricter and rejects this (which I think
is a good thing), and this is preparation for doing stronger
verification on the Rust side before we start passing data
into C.

Signed-off-by: Colin Walters <walters@verbum.org>