mkcomposefs: Reject `.` and `..` in paths
There's no good reason for us to support this; we should
expect paths to be canonicalized. In theory we *could* handle
this, but I am doubtful anyone actually relies on it.

In EROFS these are supposed to be "hard links" to the relevant
directories; the EROFS generation adds them if they don't
exist. I tried to do stronger validation at the `lcfs_node_*`
level but that is trickier.

Let's just reject at the dump file for now.

Signed-off-by: Colin Walters <walters@verbum.org>