mkcomposefs: Reject missing = in xattrs
The man page was clear that xattrs required a `=`. The code
was loose and took this case to mean an xattr with an empty
value.

Change to hard rejecting this, as it's *probably* a mistake
by anyone doing this.

This also makes silly mistakes in hand-crafting dumpfiles
a *lot* more obvious, as previously an extra `-` for example
was interpreted as an xattr with the name `-`.

Signed-off-by: Colin Walters <walters@verbum.org>