writer: Work around whiteout chardevs with nonzero size
Even though it makes no sense for character devices to have
a nonzero size, today we just faithfully copy that into the
erofs by default.

Now, a corner case rears its head when that device appears
to be a whiteout, because we end up converting it to a regular
file with a special "escaped whiteout" xattr.

When we do that, ensure we also canonialize its size to zero
because that's what the re-reading path expects.

It would honestly probably be better to just force all
character devices to have zero size, but this is a targeted fix
to clear this fuzzing case off the board.

Signed-off-by: Colin Walters <walters@verbum.org>