writer: Reject too large files
We don't currently handle non-inline files whose chunking would exceed
a single block. That limit is currently around 8 PiB which I am fairly
confident no one is going to try to put inside a composefs in
the near future.

We could lift this restriction in the future by writing multiple blocks,
but for now this fixes a crash detected by honggfuzz.

Signed-off-by: Colin Walters <walters@verbum.org>