writer: Ensure re-setting xattrs are re-validated
We did a lot of work to limit the size of xattrs, but
honggfuzz quickly figured out it was possible to bypass
this by specifying an xattr multiple times; the second
copy had special "replace existing value" code that
didn't check limits.

Fix this by always calling the "remove xattr" API when
setting (but ignoring the value). This ensures we
are consistently tracking limits.

Signed-off-by: Colin Walters <walters@verbum.org>