Add lcfs_node_set_symlink_payload
Historically we accept any arbitrary data in the dumpfile
for the file size for symlink, and end up ignoring it
ultimately when we write the EROFS. However previously
it was pretty confusing as the in-memory node data
could have a bogus size.

Add an API and use it both in the mkcomposefs path *and*
in the "reread from EROFS path" to perform more consistent
validation for symlink targets. This just wraps the
previous `lcfs_node_set_payload()` API, inheriting its
checks for length.

Signed-off-by: Colin Walters <walters@verbum.org>