tests: Move sha verification after other checks
When testing out changes to a dumpfile, today the sha check
will always fail first if I don't go and change the expected
value, but most of the time what I want to know is whether
or not it passes the *other* tests at least.

Move the checksum verification last, and also while we're
here print an "ok" value.

Signed-off-by: Colin Walters <walters@verbum.org>