Add limits and recommendations for inline content
The role of composefs is to be metadata, delegating most "regular file"
content to a backing store.

However we do support inline files (the simple case is a *zero sized*
file which happens surprisingly often); and we had an internal heuristic
of inlining files that are 64 bytes or smaller, which is less
than the SHA-256 checksum we'd need to follow them to the real
filesystem.

However...dumpfiles and `lcfs_node_set_content()` and our internal
infrastructure make no attempt to limit the size of inline files
at all.

One problem I realized here while working on fuzzing is that
there's a strict dependency between the size on the
dumpline and the inline content but many fuzzers are bad at
detecting these "strict data dependencies". So we also
change the code (only when compiled in fuzzing mode) to ignore
whatever the size is and accept what we find in the inline content.

This helped honggfuzz quickly find that we didn't handle a lot
of inline file content size at all.

Signed-off-by: Colin Walters <walters@verbum.org>