lib: Move validation out of lcfs_node_add_child()
Unfortunately at least ostree does:

node = lcfs_node_new()
lcfs_node_add_child(parent, node);
lcfs_node_set_mode(node, ...)
So we basically have a completely uninitialized node
as part of the tree.

This failed our basic "validate the mode" logic.

We can't do any validation at all in `lcfs_node_add_child()` so
move it to the first time we walk the tree as part of
`lcfs_node_write_to()`.

Also as part of fixing this: Today we don't really have coverage
of the C library as it may be used by external consumers.
Add a unit test to verify this, and we can build on this more.

Signed-off-by: Colin Walters <walters@verbum.org>