rust: Bind our default recommended max inline size
It's quite important that this gets used consistently
for interoperability - specifically, what would be really
good is to ensure that going from "tar -> composefs" has
a single well known digest. The choice of file inlining
means different composefs projects can disagree on this.

Let's encourage this to be part of the composefs standard.

Signed-off-by: Colin Walters <walters@verbum.org>