writer: Reject empty xattr names
Came out of a related discussion in ostree
https://github.com/ostreedev/ostree/pull/3346#discussion_r1866007266

Today we faithfully write this, and `fsck.erofs` doesn't seem
to complain...but such a thing makes no sense so we shouldn't
support it.

Actually digging in more here...in practice the way xattrs
work in Linux at least is they're restricted to starting with
one of the standard prefixes like `security.` or `user.`; but
enforcing/validating that is a whole other thing. For now
let's just deny this edge case.

Signed-off-by: Colin Walters <walters@verbum.org>