writer: Fix return value of lcfs_node_unset_xattr
Since the first creation of this code in
5ac1f5c4 "lib: Update xattr APIs" this function
has always returned an error code - but nothing
checked it, so it didn't matter.

Now also, currently this function cannot fail
but let's give ourselves some flexibility here;
perhaps we want to e.g. invoke `realloc`
and in that case we'd need to handle OOM.

I just noticed this while working on commit
02077e89b5b81c76e962a6643c7a7423b1f8336f
to reject empty xattr names.

Change this to return success, and also check its
return value in the set path.

Signed-off-by: Colin Walters <walters@verbum.org>