writer: make max non-inline file check constant
c1e9e9396d26 ("writer: Reject too large files") introduced a check to
ensure that we wouldn't try to write out too many null chunks for very
large external files, setting an effective limit at 8PiB (which can be
represented as 1024 chunks with the chunk format set to the maximum of
31 with a 4096 byte blocksize).

For the purposes of this check, we don't actually care what the chunking
format would be for a particular file size: we only care that it's not
going to exceed the limit.  We can calculate the limit directly using
constants and compare against that instead.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>